KJHH-1257: Add turvakielto to newly created HenkiloForceUpdateDto. Refactor henkilo duplicate handling to new DuplicateService from HenkiloService. Fix some deprecation warnings.
