+++ oppijanumerorekisteri-api/src/main/java/fi/vm/sade/oppijanumerorekisteri/dto/HenkiloForceUpdateDto.java
@@ -1,0 +1,11 @@
+package fi.vm.sade.oppijanumerorekisteri.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class HenkiloForceUpdateDto extends HenkiloUpdateDto {
+    private Boolean turvakielto;
+
+}
