+++ oppijanumerorekisteri-service/src/main/java/fi/vm/sade/oppijanumerorekisteri/services/DuplicateService.java
@@ -1,0 +1,26 @@
+package fi.vm.sade.oppijanumerorekisteri.services;
+
+import fi.vm.sade.oppijanumerorekisteri.dto.HakemusDto;
+import fi.vm.sade.oppijanumerorekisteri.dto.HenkiloDuplicateDto;
+import fi.vm.sade.oppijanumerorekisteri.dto.HenkiloDuplikaattiCriteria;
+import fi.vm.sade.oppijanumerorekisteri.models.Henkilo;
+
+import java.util.List;
+
+public interface DuplicateService {
+    List<HenkiloDuplicateDto> findDuplicates(String oid);
+
+    List<HakemusDto> getApplications(String oid);
+
+    List<HenkiloDuplicateDto> getDuplikaatit(HenkiloDuplikaattiCriteria criteria);
+
+    List<String> linkHenkilos(String henkiloOid, List<String> similarHenkiloOids);
+
+    Henkilo determineMasterHenkilo(String henkiloOid, List<String> similarHenkiloOids);
+
+    boolean hasMoreThanOneIdentifiedHenkilo(List<Henkilo> henkilos);
+
+    boolean isHenkiloIdentified(Henkilo henkilo);
+
+    void unlinkHenkilo(String oid, String slaveOid);
+}
