+++ oppijanumerorekisteri-service/src/test/java/fi/vm/sade/oppijanumerorekisteri/services/HenkiloServiceTest.java
@@ -1,7 +1,5 @@
 package fi.vm.sade.oppijanumerorekisteri.services;
 
-
-import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import com.querydsl.core.types.Predicate;
 import fi.vm.sade.kayttooikeus.dto.KayttooikeudetDto;
@@ -9,19 +7,19 @@ import fi.vm.sade.oppijanumerorekisteri.clients.AtaruClient;
 import fi.vm.sade.oppijanumerorekisteri.clients.HakuappClient;
 import fi.vm.sade.oppijanumerorekisteri.clients.KayttooikeusClient;
 import fi.vm.sade.oppijanumerorekisteri.configurations.properties.OppijanumerorekisteriProperties;
-import fi.vm.sade.oppijanumerorekisteri.mappers.OrikaConfiguration;
 import fi.vm.sade.oppijanumerorekisteri.dto.*;
 import fi.vm.sade.oppijanumerorekisteri.exceptions.NotFoundException;
+import fi.vm.sade.oppijanumerorekisteri.mappers.EntityUtils;
+import fi.vm.sade.oppijanumerorekisteri.mappers.OrikaConfiguration;
+import fi.vm.sade.oppijanumerorekisteri.models.Henkilo;
 import fi.vm.sade.oppijanumerorekisteri.models.YhteystiedotRyhma;
 import fi.vm.sade.oppijanumerorekisteri.repositories.*;
 import fi.vm.sade.oppijanumerorekisteri.repositories.criteria.HenkiloCriteria;
-import fi.vm.sade.oppijanumerorekisteri.utils.DtoUtils;
-import fi.vm.sade.oppijanumerorekisteri.mappers.EntityUtils;
-import fi.vm.sade.oppijanumerorekisteri.models.Henkilo;
 import fi.vm.sade.oppijanumerorekisteri.repositories.criteria.YhteystietoCriteria;
 import fi.vm.sade.oppijanumerorekisteri.repositories.dto.YhteystietoHakuDto;
 import fi.vm.sade.oppijanumerorekisteri.services.convert.YhteystietoConverter;
 import fi.vm.sade.oppijanumerorekisteri.services.impl.HenkiloServiceImpl;
+import fi.vm.sade.oppijanumerorekisteri.utils.DtoUtils;
 import fi.vm.sade.oppijanumerorekisteri.validators.HenkiloCreatePostValidator;
 import fi.vm.sade.oppijanumerorekisteri.validators.HenkiloUpdatePostValidator;
 import org.junit.Before;
@@ -34,19 +32,15 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.util.ReflectionTestUtils;
 
+import java.time.LocalDate;
+import java.time.Month;
 import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static fi.vm.sade.oppijanumerorekisteri.dto.YhteystietoTyyppi.*;
-import java.time.LocalDate;
-import java.time.Month;
 import static java.util.Arrays.asList;
-import static java.util.Collections.emptyList;
-import static java.util.Collections.emptySet;
-
-import static java.util.Collections.singletonList;
-import static java.util.stream.Collectors.toList;
+import static java.util.Collections.*;
 import static java.util.stream.Collectors.toSet;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
@@ -93,7 +87,7 @@ public class HenkiloServiceTest {
         this.service = spy(new HenkiloServiceImpl(this.henkiloJpaRepositoryMock, henkiloDataRepositoryMock, henkiloViiteRepositoryMock,
                 this.kielisyysRepositoryMock, this.kansalaisuusRepositoryMock, new YhteystietoConverter(), mapper, mockOidGenerator,
                 this.userDetailsHelperMock, this.permissionCheckerMock, henkiloUpdatePostValidatorMock,
-                henkiloCreatePostValidatorMock, oppijanumerorekisteriProperties, kayttooikeusClient, hakuappClient, ataruClient));
+                henkiloCreatePostValidatorMock, oppijanumerorekisteriProperties, kayttooikeusClient));
     }
 
     @Test
@@ -635,45 +629,5 @@ public class HenkiloServiceTest {
         Henkilo saved = argumentCaptor.getValue();
         assertThat(saved.getPassinumerot()).containsExactly("passi123");
     }
-
-    @Test
-    public void getHenkiloDuplicateDtoListShouldReturnApplicationsFromAtaruAndHakuApp() {
-        Henkilo henkilo1 = EntityUtils.createHenkilo("arpa noppa", "arpa", "kuutio", "hetu", "1.2.3.4.5", false, HenkiloTyyppi.OPPIJA, "fi", "FI", "2.3.34.5", new Date(), new Date(), "2.3.34.5", "arvo");
-        Henkilo henkilo2 = EntityUtils.createHenkilo("arpa2 noppa2", "arpa2", "kuutio2", "hetu2", "1.2.3.4.6", false, HenkiloTyyppi.OPPIJA, "fi", "FI", "2.3.34.5", new Date(), new Date(), "2.3.34.5", "arvo");
-        Henkilo henkilo3 = EntityUtils.createHenkilo("arpa3 noppa3", "arpa3", "kuutio3", "hetu3", "1.3.3.4.7", false, HenkiloTyyppi.OPPIJA, "fi", "FI", "2.3.34.56", new Date(), new Date(), "2.3.34.5", "arvo");
-        List<Henkilo> henkilos = Arrays.asList(henkilo1, henkilo2, henkilo3);
-        given(this.userDetailsHelperMock.getCurrentUserOid()).willReturn("1.2.3.4.5");
-        HashMap<String, Object> hakemus1 = new HashMap<>();
-        hakemus1.put("service", "ataru");
-        HashMap<String, Object> hakemus2 = new HashMap<>();
-        hakemus2.put("service", "haku-app");
-
-        HakemusDto hakemusDto1 = new HakemusDto(hakemus1);
-        HakemusDto hakemusDto2 = new HakemusDto(hakemus2);
-
-        HashMap<String, List<HakemusDto>> ataruApplications = new HashMap<String, List<HakemusDto>>() { { put("1.2.3.4.6", Arrays.asList(hakemusDto1)); }; };
-        HashMap<String, List<HakemusDto>> hakuAppApplications = new HashMap<String, List<HakemusDto>>() { { put("1.3.3.4.7", Arrays.asList(hakemusDto2)); put("1.2.3.4.6", Arrays.asList(hakemusDto2)); }; };
-
-        hakuAppApplications.put("1.3.3.4.7", Arrays.asList(hakemusDto2));
-
-        given(this.ataruClient.fetchApplicationsByOid(any())).willReturn(ataruApplications);
-        given(this.hakuappClient.fetchApplicationsByOid(any())).willReturn(hakuAppApplications);
-
-        List<HenkiloDuplicateDto> duplicates = ReflectionTestUtils.invokeMethod(service, "getHenkiloDuplicateDtoList", henkilos);
-
-        assertThat(duplicates.size()).isEqualTo(2); // filter current user
-        HenkiloDuplicateDto henkiloDuplicateDto2 = duplicates.get(0);
-        assertThat(henkiloDuplicateDto2.getHakemukset().size()).isEqualTo(2);
-        assertThat(henkiloDuplicateDto2.getHakemukset()).contains(hakemusDto1);
-        assertThat(henkiloDuplicateDto2.getHakemukset()).contains(hakemusDto2);
-
-        HenkiloDuplicateDto henkiloDuplicateDto3 = duplicates.get(1);
-        assertThat(henkiloDuplicateDto3.getHakemukset().size()).isEqualTo(1);
-        assertThat(henkiloDuplicateDto3.getHakemukset()).contains(hakemusDto2);
-
-
-    }
-
-
 
 }
